refactor: region Adapter 패키지 이동
