+++ android/app/src/main/res/layout/item_region.xml
@@ -18,7 +18,7 @@
 
 
     <TextView
-        android:id="@+id/tv_searching_region"
+        android:id="@+id/tv_region_name"
         android:layout_width="wrap_content"
         android:layout_height="wrap_content"
         android:paddingVertical="4dp"
